% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bisect.r
\docType{package}
\name{bisectr}
\alias{bisectr}
\alias{bisectr-package}
\alias{package-bisectr}
\title{bisectr package}
\description{
This package is used for creating test scripts to find bad commits with
git bisect.
For example test scripts, see \url{https://github.com/wch/bisectr}.
}

